ci: promote subgraphs from a schedule instead of waiting for the sync
Waiting for a freshly deployed version to be indexed inside the deploy job
does not hold: boson-testing-base took roughly 6 hours, the job hit its
timeout, and the 'latest' tag stayed on the previous version even though the
new one ended up indexed and healthy.

The wait is now split off from the deployment:

- the deploy job watches the new version for 5 minutes, long enough to catch
  a deployment that cannot index at all, then ends successfully and flags the
  deployed version in a SUBGRAPH_PENDING_<ENV>_<NETWORK> repository variable
- a new 'Promote deployed subgraphs' schedule checks every flagged deployment
  every 30 minutes and, once it is synced and healthy, moves the 'latest' tag
  onto it and deletes the version the tag was pointing to

Nothing holds a runner for hours any more, and no promotion can hit the 6 hour
job limit. A network is flagged, checked and cleared on its own, so a subgraph
that fails to index never affects the others.

'promote-on-goldsky.ts' makes a single check and reports 'promoted', 'pending'
or 'failed' on its 'outcome' step output, which is what clears the flag. A
transient failure - Goldsky unreachable, say - writes no outcome, so the
deployment stays flagged and is retried on the next tick.

The deploy and promote scripts share 'goldsky.ts', and the workflows share a
'setup-subgraph' composite action.
